Updated
The transparent guide and scaling box used by some objects was inheriting the check collisions setting as true. This safeguard makes sure the mold definition sets the flag to '0' to avoid collisions on the guide box. this does not affect the check collisions on the item itself, such as babylon file object or smoke.