Updated 3.2.1
If the initial start position of the avatar is over terrain, sometimes the terrain loads slower and the avatar gravity and physics setting causes the avatar to drop down and sometimes into the water below. I added a 4x1x4 transparent box right before the avatar is added so that it has a platform to stand on as the terrain loads.