updated 3.2.1
The insert value was creating a new ID field when a new value was already passed. Added if statement to check before creating new value. For Molds, this was causing additional molds to be added instead of updating (unless the web page was fully refreshed before updating a new mold).