Updated 3.2.2
Added global variable WTW.activeAnimations to assist in tracking the current user animations for movement used to report for multiplayer. In addition the setAvatarMovement included a couple references to camera that have been updated to check for current user first so that the same function can be used for positioning and setting animations for multiplayer avatars.