Updated 3.2.2
Adding Complex Parts To ActionZones - When Adding Molds (Meshes) to Action Zones like the Sliding Door and Swinging Door, if the mold is a child mold of the assembly item (common with Babylon 3D Objects), the item was not being properly added and parented to the Action Zone. After the Mold to add is picked, we now have a function called getMoldBase that you can pass the mold and get the true base of the mold before it is parented to the Action Zone. In addition, the update action zone process in the handler (and functions) is now set to just update the actionzoneid on the mold instead of re-saving the entire mold.