Updated 3.2.3
Resized Avatars in Code - The Anonymous avatars were a little small out of scale to the other avatars, so we changed the scale to be a little larger. (was .07 and now .08). In addition, the scale in code for other avatars was 0.038 scale and in the database it was only set to a precision decimal point of 2 so it was automatically rounding up to .04 scale. Now the code is set to match the database at .04 scale.