Updated 3.2.3
Resized Avatars in Code - The Anonymous avatars were a little small out of scale to the other avatars, so we changed the scale to be a little larger. (was .07 and now .08). In addition, the scale in code for other avatars was 0.038 scale and in the database it was only set to a precision decimal point of 2 so it was automatically rounding up to .04 scale. Now the code is set to match the database at .04 scale.

Enhanced Check Hovers to include Avatars - The checkHovers function along with the pluginCheckHovers hook were only processing Community, Building, and Thing molds. We enhanced the process to also include Avatars (multiplayer and MyAvatar)

Enhanced getMoldnameParts Function - Enhanced the function getMoldnameParts to include support for action zones, connecting grids, and avatars.

Set Community Mold Load Priority over Buildings and Things - In an effort to load the community molds first, added a condition to the setMoldsShown function in /core/scripts/prime/wtw_common.js to only move onto building and thing molds if all of the community molds that are needed have been processed to the queue.