Updated 3.2.3
Reformatted the Update WalkTheWeb message - The message to update WalkTheWeb was a little confusing with the date of the Latest available update and the current version date. We reformatted it to have the App Date as the latest under the App Info for update. Now there is a clear division between the current and update available.