Updated 3.2.3
Limiting running Check Action Zones - Check Action Zones was being executed every frame. I changes it to only activate when your avatar or any other avatar in your scene moves position. It uses the global variable WTW.checkZones = true to force a check and resets it to false on completion. 

Removed the FPS settings from the Check Action Zones Function - The Check Action Zones function has code to set the FPS counter, but was also setting the display, image, and style. We removed the extras to improve the FPS.

Close Menus function Updated - I Updated the closeMenus JavaScript function to close anything with the class for the left or right menu. Previously it only worked on the right. (this was prompted by the addition of the chat menu on the left)