Updated 3.2.3
Changed Start Order of Session Variables - In the file class_wtw-initsession.php, function checkHost, the order of setting session variables has been updated to support php 7.4 on Linux by initializing session parameters before starting the session. The session_start() has also been moved into this function at the appropriate order.

Correct wtw_rootpath Defined Variable - The wtw_rootpath variable was using a windows notation and now uses the forward slash in its definition. This corrects the issue of running the code from Linux.

Images Missing on 3D Objects (Linux specific) - I figured out the problem loading some images (shows as red/black checkers instead of images). When I recently moved from a Windows IIS server to a Linux server, it is now case sensitive. Turns out that some of the image files uploaded had been renamed on upload to all lowercase. For general use this would not have been an issue since it calls the image form the database which is also all lowercase in the field. But when the image is part of an assembly like a Babylon object (3D Model) then it references it in the assembly as case sensitive and the images were changed to lower case on upload. I now removed the lowercase command on the file upload on 3D Objects.