Updated 3.3.2
Remove Admin Home link if logged out - The admin home link was staying on the bottom menu bar after logout. We removed the icon and onclick events from any other related icons and text. Also, checks to see if the user has admin access to the current site before showing. This is in addition to the security check required to load a site in admin. (Contribution from TheRealFakeAdmin)