Updated 3.3.2
Major Improvement for Camera when Avatar is against a Wall - When an avatar was near a wall and would rotate, the camera preserved the distance and would show you the back side of the wall, hiding your avatar from view. With this latest advancement, we used a pick-ray to determine if a wall obstructs the camera view and automatically shorten the distance of the camera so that it doesn't lose sight of your avatar.

Set Camera Distance in the Menu - You can now set the camera distance from the avatar in the menu using a slider.