Updated 3.3.2
Updated all of the Default Avatars for Blender 2.8 - We updated the avatars to use node textures using Blender 2.8. In addition, the poly count was dramatically reduced by 40% to reduce file sizes and load times! 

We Included the Blender files for high and low poly avatars.

Added additional Color Options on Avatars - Previously we were only setting the emissive colors for the avatar parts. With the additions of node textures, there are many more options that effect colors. We now added diffuse color options to the avatar designer. In addition, diffuse also sets ambient color, while emissive also sets the specular color. You can select which color set you want using the Diffuse and Emissive buttons on the top of the color menu.