Updated 3.3.2
Released Blender 2.8 Files for Avatars! - In the near future, you will be able to upload your own avatars. To assist in helping you design, we are releasing the Blender 2.8 files for the Avatars in high and low poly count. If you use the same skeleton for your custom avatars, all of the current animations will still work!

We updated the avatars to use node textures using Blender 2.8. In addition, the poly count was dramatically reduced by 40% to reduce file sizes and load times!