Updated 3.3.2
Added additional Color Options on Avatars - Previously we were only setting the emissive colors for the avatar parts. With the additions of node textures, there are many more options that effect colors. We now added diffuse color options to the avatar designer. In addition, diffuse also sets ambient color, while emissive also sets the specular color. You can select which color set you want using the Diffuse and Emissive buttons on the top of the color menu.

Avatar Parts and materials were not Loading - On some of the Avatars, various parts (meshes) were not loading properly and the materials were also having some issues. First in the 3D Model, each part had to have its own material (copy) instead of sharing the material. This is because our designer sets the color separate for each part. Next, set the default color to black (instead of white). White causes certain materials to become polar black or white, while black absorbs the natural color of the avatar images. Finally, the meshes and materials issues were causing the system to not save your changes. This is all corrected in the designer and the main scene.