Updated 3.3.2
Added Comments to Script Files! - All of the browsing Script files now have comments. The file is commented on top to show what to find, and each function has descriptive comments. Additional comments were added as needed to describe complex functions.

Division of wtw_common.js - In order to make the code more readable and divided by function categories, we broke down the wtw_core.js file into many smaller files and added a lot of comments. New files include: wtw_utilities.js, wtw_uploads.js, wtw_login.js, wtw_dynamicscripts.js, wtw_analytics.js, molds/wtw_3dblog.js, hud/wtw_hud.js, avatars/wtw_avatarfunctions.js, and actionzones/wtw_actionzonefunctions.js.

Started basic functions for Vehicles - Created the folder scripts/vehicles and started adding functions for vehicles control and movement.

Moved Admin Menu functions to new JS File - First part of breaking down the wtw_admineditor.js file, the admin menu items are now in a new file wtw_adminmenus.js.

MoveAvatar Function moved - The Move Avatar function is now located in the newly created /core/scripts/avatars/wtw_avatarfunctions.js file. This was done to group category functions together. (FYI: it was in the /core/scripts/prime/wtw_core.js)

Transparent Material Update - Connecting grids, Action Zones, Place Holders, 3D Models, 3D Web Objects, and other complex molds use a cube to parent the parts. THese cubes were using a hidden texture and have now been updated to a standard texture with alpha set to zero (0) for the transparency.

Rename Functions to Match Naming - Some functions such as clearSoundAndLights (now disposeSoundAndLights) have been renamed to match other similar functions. Use of 'dispose' for all similar functions that dispose of items. runFunction is now executeAnimationByName, createFunctionName is now setFunctionName, checkJSFunction is now checkMoldFunctionAndExecute, soundVolumeLinear is now getSoundVolumeLinear, processCSGAction is now getMoldCSG, setCSGAction is now processCSGAction (and moldgroup was removed as a parameter).

Eliminated Obsolete Script Functions - Not all unused Scripts are obsolete, but some have been replaced by new methods. Once the new method is in place and well tested, the old Scripts can be removed. Examples include: getRotateMoldName

setShownMoldsByWeb replaces function setShownMoldsByGroup - setShownMoldsByGroup has been inactive for the past few updates. setShownMoldsByWeb was created to streamline the process with the create mold queue. This is not a change but removing the outdated process no longer in use.

Cleanup moldname parts with Function - Before the function was written, each time the name had to be parsed to get the moldgroup, molds array, index, etc... as we work with the functions we replace the references to use the WTW.getMoldnameParts(zmoldname) function.

Renamed sunLight to backLight - The direct light in the scene is the sun, the back sides of objects have a light that was named WTW.sunLight, it is now renamed as WTW.backLight so that it is not confused with the primary sun light.

Fixed typo on WTW.mainParentMold - WTW.mainParentMold had a lower case 'p' and caused it not to be set. Corrected the issue and replaced code that uses that variable.

Improved Avatar Jump Movement - Previously the avatar jump was only an animation effect. The avatar placeholder never actually left the ground. Because of this, the avatar could not jump onto higher platforms. Now, the avatar can jump and rise off the ground reasonably high levels. Like skipping a step on the stairs.

Added additional Color Options on Avatars - Previously we were only setting the emissive colors for the avatar parts. With the additions of node textures, there are many more options that effect colors. We now added diffuse color options to the avatar designer. In addition, diffuse also sets ambient color, while emissive also sets the specular color. You can select which color set you want using the Diffuse and Emissive buttons on the top of the color menu.

Avatar Parts and materials were not Loading - On some of the Avatars, various parts (meshes) were not loading properly and the materials were also having some issues. First in the 3D Model, each part had to have its own material (copy) instead of sharing the material. This is because our designer sets the color separate for each part. Next, set the default color to black (instead of white). White causes certain materials to become polar black or white, while black absorbs the natural color of the avatar images. Finally, the meshes and materials issues were causing the system to not save your changes. This is all corrected in the designer and the main scene.

Set Camera Distance in the Menu - You can now set the camera distance from the avatar in the menu using a slider.

Major Improvement for Camera when Avatar is against a Wall - When an avatar was near a wall and would rotate, the camera preserved the distance and would show you the back side of the wall, hiding your avatar from view. With this latest advancement, we used a pick-ray to determine if a wall obstructs the camera view and automatically shorten the distance of the camera so that it doesn't lose sight of your avatar.