Updated 3.3.3
Share 3D Communities, 3D Buildings, and 3D Things - For the first time since WalkTheWeb went open source, we are reinstating the ability to share 3D Communities, 3D Buildings, and 3D Things. This was previously disabled due to the additions of Babylon 3D Object Models and new graphic settings. The entire process has been reworked for security and all new complexities of 3D Websites. This is the first stage, where now we can build the library of high quality 3D items to download to your 3D Websites. We have not implemented the process to search and download those items yet (until we have enough to replace the former repository).