Updated 3.3.4
Fixed Right Click on 3D Thing with 3D Models - When you right click in Admin mode on a 3D Model that is part of a 3D Thing, it was sending you to the wrong form (mold edit instead of connecting grid form) even though the 3D Model is not part of the item being edited. Corrected the process to select 3D Things and 3D Buildings in 3D Communities by editing their connecting grid.