Updated 3.3.4
New Color Selector Wheel Process - New Color Selector Wheel Process - When editing a mold, if the covering (texture selection) sets a color, you can now set diffuse, emissive, specular, and ambient colors for a mold with a color wheel selector and/or html color codes. The color wheel is larger and only a single wheel shown at a time. Just select the textbox and the color wheel will automatically open. It closes when focus changes to another field or save, delete, or cancel the mold changes.