Updated 3.3.4
Created a New 3D Webs Repository - Created a new 3D Webs (3D Communities, 3D Buildings, and 3D Things) repository with all of the latest functionality. Some of the new features are the inclusion of 3D Models, avatar animations, and scripts. This new functionality can be found in the new WalkTheWeb install or the WalkTheWeb downloads in the media library.