Updated 3.3.4
Added Avatar List to Admin Menu - Added an Admin Mode menu item for Avatars with a submenu item for Avatars List. In the near future, this will be used for admin to set the default Avatars, colors, and animations.

Added avatargroups Table - With the new Admin Mode menu options, we added a table called avatargroups that lists the categories of avatars in the default avatars. This is the first step towards being able to assign specific avatar groups to 3D Community Scenes, like medieval, space, or zombies, where each visitor must select an appropriate avatar for the 3D Scene.

Added Avatar Groups to Admin Menu - Added an Admin Mode submenu item for Avatar Groups. In the near future, this will be used for admin to group avatars and assign them to 3D Community Scenes.