Updated 3.3.4
Connecting Grid related MoldName Update - When the Mold uses an alternate Load Zone from the parent, the mold name was wrongly named using the parent connecting grid. Corrected the issue so that the mold name properly represents the connecting grid it comes from instead of the top parent.

Added pastuploadobjectid and  pastobjectanimationid to tables - Used by the shared downloads, the pastuploadobjectid and pastobjectanimationid  allows us to track items from a particular table that have been downloaded, since the id are made to be unique and often gets a new random value. This fields have been added to the tables class in php (pastuploadobjectid to uploadobjects table and pastobjectanimationid  to uploadobjectanimations table).

Thing within a Building wrong Location - When you have a 3D Community with a 3D Building that has a 3D Thing in the 3D Building (3 levels of parenting), the 3D Thing was not setting the proper location and rotation. Turns out that the connecting grid was not being loaded to get the position and rotation. The /connect/connectinggrids.php file was corrected with an 'if' statement instead of an 'else if' needed to process the second connecting grid in a single record.