Updated 3.3.4
Thing within a Building wrong Location - When you have a 3D Community with a 3D Building that has a 3D Thing in the 3D Building (3 levels of parenting), the 3D Thing was not setting the proper location and rotation. Turns out that the connecting grid was not being loaded to get the position and rotation. The /connect/connectinggrids.php file was corrected with an 'if' statement instead of an 'else if' needed to process the second connecting grid in a single record.