Updated 3.3.4
Created a New 3D Webs Repository - Created a new 3D Webs (3D Communities, 3D Buildings, and 3D Things) repository with all of the latest functionality. Some of the new features are the inclusion of 3D Models, avatar animations, and scripts. This new functionality can be found in the new WalkTheWeb install or the WalkTheWeb downloads in the media library. 

Added pastuserid to Users table - Used by the shared downloads, the pastuserid allows us to track items from a particular user that have been downloaded, since the userid is made to be unique and often gets a new random value. This field has been added to the tables class in php.

Created getNewKey PHP function - The getNewKey PHP function has been added to the wtwdb class. The function lets you pass the tablename (without prefix), key field name, and possible first key value to test and the function will use the default, then random string values until it returns a unique value. This function has been extended to the handlers and connect classes.

Created getIDByPastID PHP function - Mostly used during import (shared downloads) process, if the key id of a table was changed due to duplicates, the old value is stored in the past fieldname field. This function lets you retrieve the new value based on the past field.  Dates are used to select the latest where duplicates are found. Duplicates exist most often when the same 3D Web is downloaded multiple times. This function was added to the wtwdb class and has been extended to the handlers and connect classes.

Added Profile Fields to Users table - Added profile fields for firstname, lastname, gender, and dateofbirth (DOB) into the users table. This will extend the basic profile for the users.

Added PHP function for Field Exists in a Table - Added a PHP function to determine if a field (column) exists in a given table. The function was added to the wtwdb class and extended to plugins, handlers, and connect classes.

Added avatargroups Table - With the new Admin Mode menu options, we added a table called avatargroups that lists the categories of avatars in the default avatars. This is the first step towards being able to assign specific avatar groups to 3D Community Scenes, like medieval, space, or zombies, where each visitor must select an appropriate avatar for the 3D Scene.