Updated 3.3.4
Default Texture Update - In code, the default texture on many new objects was the stock image 'lightgray-512x447.jpg'. It has been updated to be a 512x512 image 'lightgray-512x512.jpg'.

Thing within a Building wrong Location - When you have a 3D Community with a 3D Building that has a 3D Thing in the 3D Building (3 levels of parenting), the 3D Thing was not setting the proper location and rotation. Turns out that the connecting grid was not being loaded to get the position and rotation. The /connect/connectinggrids.php file was corrected with an 'if' statement instead of an 'else if' needed to process the second connecting grid in a single record. 

Web Alias Editing fixed - In admin mode, web aliases are set to map paths directly to specific domains, communities, buildings, and things. When editing a web alias, the dropdown menus were not set to the previous selected item. This has been corrected to show the last setting when editing. In addition, we set the focus to the form so that long lists of web aliases scroll back to the top when you click edit.

Update user was dropping Email - When admin updates a user, the save process had a mislabeled email (needed to be useremail) when passing the value. Made the correction and is now saving the email address.

Created a New 3D Webs Repository - Created a new 3D Webs (3D Communities, 3D Buildings, and 3D Things) repository with all of the latest functionality. Some of the new features are the inclusion of 3D Models, avatar animations, and scripts. This new functionality can be found in the new WalkTheWeb install or the WalkTheWeb downloads in the media library. 

Users table key Field Change - In the users table, the primary key was set to the username, yet the values used as the foreign key was userid. For consistency and proper setup, the primary key is now userid and username is set to not null. wtwtables class has been updated.

Adjusted Admin Menu Scroll Height - Adjusted the Admin Menu Scroll height so that it no longer hides part of the bottom buttons under the Quick Editor Settings.

New Color Selector Wheel Process - When editing a mold, if the covering (texture selection) sets a color, you can now set diffuse, emissive, specular, and ambient colors for a mold with a color wheel selector and/or html color codes. The color wheel is larger and only a single wheel shown at a time. Just select the textbox and the color wheel will automatically open. It closes when focus changes to another field or save, delete, or cancel the mold changes.

Eliminated the use of RGB Color Fields - The following fields have been eliminated from use and therefore deleted from the code: wtw_tspecularcolorr, wtw_tspecularcolorg, wtw_tspecularcolorb, wtw_tdiffusecolorr, wtw_tdiffusecolorg, wtw_tdiffusecolorb, wtw_temissivecolorr, wtw_temissivecolorg, wtw_temissivecolorb. The new version uses hex code for wtw_tmolddiffusecolor, wtw_tmoldemissivecolor, wtw_tmoldspecularcolor, wtw_tmoldambientcolor.

Added Hex color Updates to Table Values - First time you go to admin mode, the hex colors will be populated based on the values of rgb colors in the tables. This applies to communitymolds, buildingmolds, thingmolds, and useravatarcolors. To perform this task, the PHP function getHexFromRGB was added to the wtwdb_class and updateColorsHex was added to the wtwtables class.

Replaced MoldGroup variable for WebType - WebType and MoldGroup variables were interchangable for meaning. Therefore, we eliminated the use of moldgroup by replacing the instances with WebType. This includes the field name reference in the Scripts table.

Added hexToRGB and rgbToHex functions - Added hexToRGB and rgbToHex functions to the core/scripts/prime/wtw_utilities.js. This will make it easier to convert back and forth between rgb and hex.

Added hex color fields to tables - Added hex color fields for diffusecolor, specularcolor, emissivecolor, and ambientcolor as varchar(7) to the tables: communitymolds, buildingmolds, thingmolds, and useravatarcolors. This will assist in phasing out the storing of red, green, blue number colors individually.

Fixed Uploads foreign key assignments for Images - Fixed Uploads foreign key assignments for Images. The previous method used the database to confirm the identity key. Now it creates all of the required keys and adds them to each record on insert avoiding update errors.

Removed Displayname from User Authorizations - Removed the displayname field from the userauthorizations table. It was depreciated and no longer in use.