Updated 3.3.4
Added hex color fields to tables - Added hex color fields for diffusecolor, specularcolor, emissivecolor, and ambientcolor as varchar(7) to the tables: communitymolds, buildingmolds, thingmolds, and useravatarcolors. This will assist in phasing out the storing of red, green, blue number colors individually.

Replaced MoldGroup variable for WebType - WebType and MoldGroup variables were interchangable for meaning. Therefore, we eliminated the use of moldgroup by replacing the instances with WebType. This includes the field name reference in the Scripts table.

Variable Renaming (also effects Plugins) - We are continuing our efforts to be consistent on variable naming. Completed a lot more local variables within JavaScript functions to start with the letter z. This makes it easier to identify local vs global variables at a glance. The plugin hooks also had the variables passed to them all renamed to start with z.

Eliminated the use of RGB Color Fields - The following fields have been eliminated from use and therefore deleted from the code: wtw_tspecularcolorr, wtw_tspecularcolorg, wtw_tspecularcolorb, wtw_tdiffusecolorr, wtw_tdiffusecolorg, wtw_tdiffusecolorb, wtw_temissivecolorr, wtw_temissivecolorg, wtw_temissivecolorb. The new version uses hex code for wtw_tmolddiffusecolor, wtw_tmoldemissivecolor, wtw_tmoldspecularcolor, wtw_tmoldambientcolor.

Reset new Molds Y value - Originally the Y value (up) was set based on the avatar or camera location with an offset. Now that the avatars position is redesigned, we still use avatar position or camera, but needed to update the offset for better placement.

Deleted table fields for RGB Colors for Molds (optional) - With the new change in HTML color codes instead of RGB on tables communitymolds, buildingmolds, and thingmolds, once the update has complete if you check that the colors are working (populated in the edit mold form - right click a basic mold in the scene), then you can manually delete the table fields for (notice they end in RGB - not to delete the new fields without the R, G,or B reference) specularcolorr, specularcolorg, specularcolorb, diffusecolorr, diffusecolorg, diffusecolorb, emissivecolorr, emissivecolorg, and emissivecolorb (only in the 3 molds tables). 

ObjectFolder and ObjectFile removed from Mold tables - The fileds objectfolder and objectfile have been retired from the tables: buildingmolds, communitymolds, and thingmolds. Optionally, the fields can be manually deleted from the molds tables. It has been depreciated to the values in the related objectuploads table. Note that this does not effect other tables like avatar and animation related folders and files.

Fixed Extended Ground Texture Change and Sun Spot - The extended ground was not updating when changed. Fixed the preview thumbnail when selected and the save JavaScript was trying to use the getJSON function instead of the correct postJSON function. In addition, adjusted the specular color of extended ground to tone down the sun spot.

Connecting Grid related MoldName Update - When the Mold uses an alternate Load Zone from the parent, the mold name was wrongly named using the parent connecting grid. Corrected the issue so that the mold name properly represents the connecting grid it comes from instead of the top parent.

Corrected Saving Names, AltTags, and Descriptions - The Handlers were not consistent with how data was passed from base64 or html encoded, causing some entries to be dropped. Now JavaScript is passing those values using base64 and PHP decodes the base 64 on read in the handler.

Thing within a Building wrong Location - When you have a 3D Community with a 3D Building that has a 3D Thing in the 3D Building (3 levels of parenting), the 3D Thing was not setting the proper location and rotation. Turns out that the connecting grid was not being loaded to get the position and rotation. The /connect/connectinggrids.php file was corrected with an 'if' statement instead of an 'else if' needed to process the second connecting grid in a single record. 

Default Texture Update - In code, the default texture on many new objects was the stock image 'lightgray-512x447.jpg'. It has been updated to be a 512x512 image 'lightgray-512x512.jpg'.

Added Keystroke Filter for safetext - Keyboard filters assist with front-end validation of keys available for each textbox. safetext was added to accommodate text descriptions with punctuation with the exception of greater than and less than <> symbols. 

Sync max length of Fields with Database - Some of the fields in the user table is set to 255 characters while the forms in the code used maxlength 256. Adjusted the forms maxlength to match the database.

Added Profile View and Edit Forms - Added forms to view and edit the logged in local profile. It includes: Display Name, Email, First Name, Last Name, Gender, and DOB. 

Combined Select My Avatar with Quick Start Avatars - In an attempt to lessen the number of clicks to get started, we added a new login menu select avatar page that combines the Select My Avatar (previous avatars created) with Quick Start Avatars (list of available avatars) in one form. 

Browse Menu Link for Profile Redirected to New Profile Form - On the bottom browse menu after login, the user name and profile icon link to the profile form. This link has been updated to the new profile form.