Updated 3.3.4
Avatar Start Platform Increased - When the avatar enters the 3D Scene, if the avatar loads faster than the ground beneath it, the avatar can drop to the extended ground level. Because of this, we added an invisible box under the avatar to hold the avatar up before the ground renders. The previous box was small and occasionally the avatar would move forward and fall off. We modified the box to be much larger and stay for 10 seconds before removing it.