Updated 3.3.4
Avatar Start Platform added after Avatar Selection - The Start Platform was used when the 3D Scene loads, but if you selected your avatar after the 10 second timer, the start platform was not there when the avatar loaded. We added the start platform to the reload process so that any new or changed avatar has the safety platform under the avatar for the first 10 seconds.