Updated 3.3.4
Change Avatar Colors from RGB to Hex - Changed the avatar color database fields to use Hex values instead of RGB saved in the database. Everything now supports diffuse, specular, emissive, and ambient except for the user interface, which will be updated soon. This affected the JSON fetched data and save procedures, local and global avatars, and 3D Internet plugin in development. 

Display Name Consistency - The display name (shown on the menu bar) is now the Display name of your chosen avatar. If one does not exist or is blank, it defaults to the display name in your user profile, with a fallback to the username, then finally the first half of your email address before the @.