Updated 3.3.4
Some New Action Zones Not Showing - When you add a new Swinging Door action zone, it was not visible on the 3D Scene until you saved it and edited it again. Fixed the visibility issue and also cleaned up the code for action zones by starting local variables with z and removing unused variables.