Updated 3.3.4
Admin Menu Updates - Updated the Admin Menu by removing the 'My' from the 3D Items, renamed Avatars to 3D Avatars, and reordered the listings.

Set initial Avatar Colors on New Avatars - Some new avatars were starting all white (Default Female and Male). Created a table for default avatar colors, added default data for each avatar, and added initial color assignments to the new avatar selection process.

Added Parameters to Admin Users List - With the addition of the new download repository, it was also necessary to add query parameters to not include users with a pastuserid, therefore only showing Local and Global users that have actually visited your 3D websites.

Added Parameters to Uploads - In the media library, we added query parameters to not include WalkTheWeb download uploads in the lists of my images and stock images. This prevents unnecessary clutter. The images will still show on the 3D Community, 3D Building, or 3D Thing tag when the image is in the currently loaded 3D Webs object being edited.

Stock image path correction - The stock image for grass.jpg was set to the wrong file path as grasslawn.jpg. This also effects the 512x512 version of the file path. The path has been corrected in the wtw_tables PHP class.

Local Login Parameters Added - With the addition of the new download repository, it has the ability to add users in namesake to the users table in order to identify who created or updated an object last. These users will always have a pastuserid listed, which has now been added to the local login query parameters. 

Change Avatar Colors from RGB to Hex - Changed the avatar color database fields to use Hex values instead of RGB saved in the database. Everything now supports diffuse, specular, emissive, and ambient except for the user interface, which will be updated soon. This affected the JSON fetched data and save procedures, local and global avatars, and 3D Internet plugin in development. 

Automated Activating Default 3D Plugins - During a new install there are a couple of 3D Plugins that are useful from the start. These default 3D Plugins are now set to active on install and include: wtw-avatars and wtw-shopping.

Pluginid field removed from plugins table - The Plugin Name is a unique identifier since it is also the folder for the plugin. Therefore, there is no need for an additional pluginid. The field has been removed from the code (was not in use anywhere).

New Install was missing some Initial Data - During a new install, some of the tables were not receiving the initial data. Corrections to fields and data have been made to properly format the data import.

Added Frame information to initial Avatars data - The Avatars table supported startframe and endframe fields for the initial included wait animation but the initial data did not include the values. Corrected and added the initial values.

Added Parent Download Tracking fields - Added downloadparentwebid and downloadparentwebtype fields to tables: communities, buildings, and things in order to assist with tracking supplemental downloads. For example, if you download a 3D Community that has a building already on it, the building record in the table will show the communityid as the downloadparentwebid and 'community' as the downloadparentwebtype.

Added First Building Position and Rotation to Communities - The automated install and quick start processes include the ability to add the first 3D Building to your selected 3D Community Scene. We added position and rotation fields to the Community table (buildingpositionx, buildingpositiony, buildingpositionz, buildingrotationy) to accommodate the automated processes.

Added avatargroups Table - With the new Admin Mode menu options, we added a table called avatargroups that lists the categories of avatars in the default avatars. This is the first step towards being able to assign specific avatar groups to 3D Community Scenes, like medieval, space, or zombies, where each visitor must select an appropriate avatar for the 3D Scene.

Added PHP function for Field Exists in a Table - Added a PHP function to determine if a field (column) exists in a given table. The function was added to the wtwdb class and extended to plugins, handlers, and connect classes.

Added avatarcolors Table - With the new Admin Mode menu options, we added a table called avatarcolors that will hold the default colors for avatar parts (meshes). This is the counter-part of the useravatarcolors table that holds the custom colors chosen for the avatar parts (meshes) per individual users.

Added Avatar List to Admin Menu - Added an Admin Mode menu item for Avatars with a submenu item for Avatars List. In the near future, this will be used for admin to set the default Avatars, colors, and animations. 

Added Profile Fields to Users table - Added profile fields for firstname, lastname, gender, and dateofbirth (DOB) into the users table. This will extend the basic profile for the users.

Update user was dropping Email - When admin updates a user, the save process had a mislabeled email (needed to be useremail) when passing the value. Made the correction and is now saving the email address.