Updated 3.3.4
Avatar Designer Window Adjustments - When editing the avatar, the window was wrapping under the bottom menu and the contents of the window did not fill the lower part of the window. Adjusted the vertical size of the window and set the background to black for the designer window only. Set the background opacity back to .2 when it closes and reopens.