Updated 3.3.4
Display Name Consistency - The display name (shown on the menu bar) is now the Display name of your chosen avatar. If one does not exist or is blank, it defaults to the display name in your user profile, with a fallback to the username, then finally the first half of your email address before the @.

Set initial Avatar Colors on New Avatars - Some new avatars were starting all white (Default Female and Male). Created a table for default avatar colors, added default data for each avatar, and added initial color assignments to the new avatar selection process.

Camera to Avatar Transitions - Changed the initial camera angle to match the 90 degree offset (avatars are 90 degrees from the avatar base to make the animations relative). Then when the avatar enters the 3D Scene, we match the avatar rotation Y angle to the current camera rotation Y for a much smoother transition.

Local Login Parameters Added - With the addition of the new download repository, it has the ability to add users in namesake to the users table in order to identify who created or updated an object last. These users will always have a pastuserid listed, which has now been added to the local login query parameters. 

Change Avatar Colors from RGB to Hex - Changed the avatar color database fields to use Hex values instead of RGB saved in the database. Everything now supports diffuse, specular, emissive, and ambient except for the user interface, which will be updated soon. This affected the JSON fetched data and save procedures, local and global avatars, and 3D Internet plugin in development. 

Temporary Avatar Refresh for Colors - Previously, when an avatar colors change in the editor, the 3D Scene would update the colors without having to fully reload the avatar. Currently, the materials refresh is not working so we enabled the full reload as a temporary fix. In an upcoming update we will be correcting this issue as we also switch to storing the hex color values instead of decimal RGB colors. 

Enabled same domain for onMessage Events - The event listener for onMessage that sends messages from approved iframes back to the main parent window of WalkTheWeb now allows messages from the originating domain. Only certain commands are allowed like closing the iframe window.

Added First Building Position and Rotation to Communities - The automated install and quick start processes include the ability to add the first 3D Building to your selected 3D Community Scene. We added position and rotation fields to the Community table (buildingpositionx, buildingpositiony, buildingpositionz, buildingrotationy) to accommodate the automated processes.